+++ src/plugins/vis_builder/public/application/components/workspace.tsx
@@ -4,7 +4,7 @@
  */
 
 import { i18n } from '@osd/i18n';
-import { EuiEmptyPrompt, EuiFlexGroup, EuiFlexItem, EuiIcon, EuiPanel } from '@elastic/eui';
+import { EuiEmptyPrompt, EuiFlexItem, EuiIcon, EuiPanel } from '@elastic/eui';
 import React, { useState, useMemo, useEffect, useLayoutEffect } from 'react';
 import { useOpenSearchDashboards } from '../../../../opensearch_dashboards_react/public';
 import { IExpressionLoaderParams } from '../../../../expressions/public';
@@ -18,7 +18,6 @@ import hand_field from '../../assets/hand_field.svg';
 import fields_bg from '../../assets/fields_bg.svg';
 
 import './workspace.scss';
-import { ExperimentalInfo } from './experimental_info';
 import { handleVisEvent } from '../utils/handle_vis_event';
 
 export const WorkspaceUI = () => {
@@ -105,11 +104,6 @@ export const WorkspaceUI = () => {
 
   return (
     <section className="vbWorkspace">
-      <EuiFlexGroup className="vbCanvasControls">
-        <EuiFlexItem>
-          <ExperimentalInfo />
-        </EuiFlexItem>
-      </EuiFlexGroup>
       <EuiPanel className="vbCanvas" data-test-subj="visualizationLoader">
         {expression ? (
           <ReactExpressionRenderer