+++ src/plugins/vis_builder/public/plugin.test.ts
@@ -39,7 +39,6 @@ describe('VisBuilderPlugin', () => {
           title: PLUGIN_NAME,
           aliasPath: '#/',
           aliasApp: PLUGIN_ID,
-          stage: 'experimental',
         })
       );
     });