+++ src/plugins/vis_builder/public/plugin.ts
@@ -188,7 +188,6 @@ export class VisBuilderPlugin
         defaultMessage: 'Create visualizations using the new VisBuilder',
       }),
       icon: 'visBuilder',
-      stage: 'experimental',
       aliasApp: PLUGIN_ID,
       aliasPath: '#/',
       appExtensions: {
@@ -201,7 +200,6 @@ export class VisBuilderPlugin
             icon: 'visBuilder',
             id,
             savedObjectType: VISBUILDER_SAVED_OBJECT,
-            stage: 'experimental',
             title: attributes?.title,
             typeTitle: VIS_BUILDER_CHART_TYPE,
             updated_at: updatedAt,