+++ src/plugins/vis_builder/public/services/type_service/visualization_type.tsx
@@ -14,7 +14,7 @@ export class VisualizationType implements IVisualizationType {
   public readonly title: string;
   public readonly description: string;
   public readonly icon: IconType;
-  public readonly stage: 'experimental' | 'production';
+  public readonly stage: 'production';
   public readonly ui: IVisualizationType['ui'];
   public readonly toExpression: (
     state: RenderState,